🔨 refactor: LetterWrite 바텀시트 분리
